Renames tests/airflow_client.py to tests/test_airflow_client.py to fix import errors.
