+++ tests/apps/vendor_app/test_vendor_management.py
@@ -7,7 +7,7 @@ from libsys_airflow.plugins.vendor.models import (
     Vendor,
     VendorInterface,
 )
-from tests.airflow_client import test_airflow_client  # noqa: F401
+from tests.test_airflow_client import test_airflow_client  # noqa: F401
 
 now = datetime.now(timezone.utc)
